# Do not re-encode head chunk for ChunkQuerier