--- a/tsdb/test.txt
+++ b/tsdb/test.txt
@@ -1,0 +1,1 @@
+make: Nothing to be done for `test'.