Currently, we re-encode the head chunk (by iterating through all samples and re-compressing them into a chunk) while using ChunkQuerier to avoid races with appends. It is much more efficient to just take a copy of chunk bytes.

```
benchmark                        old ns/op     new ns/op     delta
BenchmarkHeadChunkQuerier-10     575568        199947        -65.26%

benchmark                        old allocs     new allocs     delta
BenchmarkHeadChunkQuerier-10     2720           2320           -14.71%

benchmark                        old bytes     new bytes     delta
BenchmarkHeadChunkQuerier-10     256892        215310        -16.19%
```

These are the benchmark results for querying samples; unaffected. Sometimes delta of running time is ~0%, but many times there is an improvement (which should not be the case actually).
```
benchmark                   old ns/op     new ns/op     delta
BenchmarkHeadQuerier-10     1239021       1216053       -1.85%

benchmark                   old allocs     new allocs     delta
BenchmarkHeadQuerier-10     2820           2720           -3.55%

benchmark                   old bytes     new bytes     delta
BenchmarkHeadQuerier-10     213688        202484        -5.24%
```