@codesome 2023-03-02
> Frankly I'm a bit lost in populateWithDelChunkSeriesIterator, populateWithDelSeriesIterator, populateWithDelGenericSeriesIterator - I have very little idea what each does and when they are used.

I feel you, I also get lost everytime. From the looks of it `populateWithDelGenericSeriesIterator` is generic implementation that helps populate the chunks while taking care of deletions (i.e. tombstones). `populateWithDelChunkSeriesIterator` and `populateWithDelSeriesIterator` use that genetic iterator to finally provide chunks and samples respectively. Some interface/plugging magic going on in the upper chain - this refactor was done to make life easy for downstream consumers and probably also helps Prometheus in some ways.